% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabellr.R
\name{fparse_jsontab_scb}
\alias{fparse_jsontab_scb}
\title{parse json respone from scb.se#'}
\usage{
fparse_jsontab_scb(ldatajson)
}
\arguments{
\item{ldatajson}{json response (content) from scb.se api}
}
\value{
table with correctly formated columns
}
\description{
parse json respone from scb.se#'
}
\references{
copied in its entiery from https://dittochdata.netlify.app/2018/06/10/ladda-hem-data-fr%C3%A5n-scb-del-2-anv%C3%A4nda-scb-s-api/
}
\author{
Christian Lindell
}
