% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabellr.R
\name{scb_meta_unpack}
\alias{scb_meta_unpack}
\title{unpack meta-data from scb.se}
\usage{
scb_meta_unpack(n)
}
\arguments{
\item{n}{number of fields to unpack}
}
\value{
list with lookups for table dimensions
}
\description{
unpack meta-data from scb.se
}
\details{
"tabellinnehall" and "vartannat ar" excluded - no meta needed for these
}
