% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabellr.R
\name{scb_results}
\alias{scb_results}
\title{post, extract content and parse}
\usage{
scb_results()
}
\value{
correctly formated table
}
\description{
makes the post, extracts payload and parses with fparse_jsontab_scb
}
