% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabellr.R
\name{scb_tabell}
\alias{scb_tabell}
\title{create table with data from scb.se}
\usage{
scb_tabell(path_to = p1)
}
\value{
df with data from selected indicator on scb.se
}
\description{
top-level, creats table

inner function loops over all lookups in scb_metadata.
}
